python/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch>=2.6.0
transformers>=4.50.0
accelerate
pillow
requests
sentencepiece
protobuf
bitsandbytes
fastapi
uvicorn
python-multipart

